Update to Exchange Contract and Swap Contract
Fix the SwapFee temporarily until the LP gets deployed and then that part of the contract won't affect anything.
